+++ src/blog_api/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
-
+from .models import Category, Post
 # Register your models here.
+
+admin.site.register(Category)
+admin.site.register(Post)
