--- src/user/models.py
@@ -1,24 +1,0 @@
-from typing import Optional
-from django.db import models
-from django.db.models.fields import BLANK_CHOICE_DASH
-from django.contrib.auth.models import User  # import django user model
-# Create your models here.
-
-class Category(models.Model):
-    name = models.CharField(max_length=100)
-
-
-class Post(models.Model):
-    OPTIONS = (
-        ( "d", "Draft" ),
-        ( "p", "Publised" )
-    )
-    title = models.CharField(max_length=100)
-    content = models.TextField()
-    image = models.ImageField()
-    category = models.ForeignKey(Category, on_delete=models.PROTECT)  # Categoriye ait olan postların kategori silinince silinmesini engellemek için PROTECT
-    publish_date = models.DateTimeField(auto_now_add=True)
-    last_updated = models.DateTimeField(auto_now=True)
-    author = models.ForeignKey(User, on_delete=models.CASCADE)  # user silinince POST larında silinmesi için CASCADE
-    status = models.CharField(max_length=10, choices=OPTIONS, default="d") 
-    slug = models.SlugField(blank=True, unique=True)
